backend/requirements.txt
# Backend Dependencies for RAG Demo Chatbot

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Vector Database
qdrant-client==1.12.1

# AI/LLM
google-generativeai==0.8.3

# Document Processing
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# Job Scheduling
APScheduler==3.10.4

# Configuration
python-dotenv==1.0.0
pydantic==2.10.4
pydantic-settings==2.7.1

# CORS
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
